+++ test/tls_certificate_check_SUITE.erl
@@ -65,7 +65,7 @@ init_per_suite(Config) ->
     Config.
 
 end_per_suite(_Config) ->
-    ok.
+    ok = application:stop(tls_certificate_check).
 
 init_per_group(GroupName, Config) ->
     case GroupName of